requirements.txt
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.3
django-crispy-forms==1.7.2
django-import-export==1.1.0
django-widget-tweaks==1.4.3
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.10
pipdeptree==0.13.2
psycopg2-binary==2.7.6.1
pytz==2018.7
PyYAML==3.13
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==4.1.2
xlrd==1.1.0
xlwt==1.3.0
